Move compute_price and compute_tax generic methods to sale_base
