+++ anyblok_sale/bloks/sale_base/__init__.py
@@ -23,9 +23,9 @@ class SaleBaseBlok(Blok):
 
     @classmethod
     def import_declaration_module(cls):
-        from . import model # noqa
+        from . import base # noqa
 
     @classmethod
     def reload_declaration_module(cls, reload):
-        from . import model
-        reload(model)
+        from . import base
+        reload(base)
