--- anyblok_sale/bloks/sale_base/model.py
@@ -1,14 +1,0 @@
-# This file is a part of the AnyBlok / Sale project
-#
-#    Copyright (C) 2018 Franck Bret <franckbret@gmail.com>
-#
-# This Source Code Form is subject to the terms of the Mozilla Public License,
-# v. 2.0. If a copy of the MPL was not distributed with this file,You can
-# obtain one at http://mozilla.org/MPL/2.0/.
-# -*- coding: utf-8 -*-
-from anyblok import Declarations
-
-
-@Declarations.register(Declarations.Model)
-class Sale:
-    """Namespace for Sale related models"""
